sql实现，mysql

+++ domain/src/main/java/com/kingtang/area/sf/repository/StateSeriesRepository.java
@@ -7,5 +7,5 @@ import java.math.BigDecimal;
 public interface StateSeriesRepository {
     void add(StateSeries stateSeries);
 
-    BigDecimal getNewId();
+    Long getNewId();
 }
